requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.11
numpy>=1.24.0
pathlib2>=2.3.7
psutil>=5.9.0
